backend/requirements.txt

Flask==3.1.0
gunicorn==23.0.0
flask-cors==3.0.10
python-dotenv==1.0.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.27
werkzeug==3.0.3
